not_logged_in_search-for-songs_search-for-users_users-show

+++ resources/views/favorite/favorite_button.blade.php
@@ -1,13 +1,23 @@
-@if (Auth::user()->is_favoriting($song->id))
-    {!! Form::open(["route" => ["favorites.unfavorite", $song->id], "method" => "delete"]) !!}
-        <i class="fas fa-star favorite-orange"></i>
-        {{ count($song->favorite_users) }}
-        {!! Form::submit("登録済み", ["class" => "btn btn-warning btn-sm btn-unfavorite"]) !!}
-    {!! Form::close() !!}
+<!--ログインしているとき-->
+@if(Auth::check())
+    @if (Auth::user()->is_favoriting($song->id))
+        {!! Form::open(["route" => ["favorites.unfavorite", $song->id], "method" => "delete"]) !!}
+            <i class="fas fa-star favorite-orange"></i>
+            {{ count($song->favorite_users) }}
+            {!! Form::submit("登録済み", ["class" => "btn btn-warning btn-sm btn-unfavorite"]) !!}
+        {!! Form::close() !!}
+    @else
+        {!! Form::open(["route" => ["favorites.favorite", $song->id]]) !!}
+            <i class="fas fa-star"></i>
+            {{ count($song->favorite_users) }}
+            {!! Form::submit("登録する", ["class" => "btn btn-default btn-sm btn-favorite"]) !!}
+        {!! Form::close() !!}
+    @endif
+    
+<!--ログインしていない時-->
 @else
-    {!! Form::open(["route" => ["favorites.favorite", $song->id]]) !!}
+    <div>
         <i class="fas fa-star"></i>
         {{ count($song->favorite_users) }}
-        {!! Form::submit("登録する", ["class" => "btn btn-default btn-sm btn-favorite"]) !!}
-    {!! Form::close() !!}
+    </div>
 @endif
